Fix permissions only on our installed directory, not all of /opt
this task:
 - name: Change wildfly installation permissions
   file: path={{ wildfly_install_dir }} owner=wildfly group=wildfly mode=0750

is changing the directory permissions on wildfly_install_dir (= /opt),
not just on the directory tree that we created (/opt/wildfly-*).  This
leads to all of /opt being incorrectly permissioned, affecting
anything else that may be installed into /opt.

The fix is to use wildfly_dir instead of wildfly_install_dir in this
task, and in the final task of the play which does likewise.

file: path={{ wildfly_dir }} owner=wildfly group=wildfly mode=0750
          recurse=yes