Fixes #8
A systemd definition requires the identification of a given Service, whether it is a service or a socket, etc. This bugfix fixes #8. But there are more things to do. The init script allowed for a lot more options, which need to be ported into the systemd script. I think the easiest would be to call the init script with the start option from the service definition file. Let me know what you think.